Added a construct  function and suport for file upload Now User will not be allowed to modified posts from other users.

+++ app/Http/Controllers/PostsController.php
@@ -4,8 +4,19 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Post;
+use Illuminate\Support\Facades\Storage;
 class PostsController extends Controller
 {
+        /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -37,14 +48,31 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
 
+        //Handle File Upload
+        if($request->hasFile('cover_image')){
+            //Get filename with extencion.
+            $fileNameWithExt = $request ->file('cover_image')-> getClientOriginalName();
+            //Get just fileName
+            $fileName = pathInfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get just Ext
+            $fileExt = $request->file('cover_image')->getClientOriginalExtension();
+            //Filename to Store
+            $fileNameToStore = $fileName.'_'.time().'.'.$fileExt;
+            //Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noImage.jpg';
+        }
         //Create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
         return redirect('/posts')->with('success', 'Post Created');
     }
@@ -70,6 +98,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $posts = Post::find($id);
+        //Chek for correct user
+        if(auth()->user()->id != $posts->user_id) {
+            return redirect('/posts')->with('error', 'Unathorized');
+        }
+
         return view('posts.edit')->with('post', $posts);
     }
 
@@ -86,11 +119,26 @@ class PostsController extends Controller
             'title' => 'required',
             'body' => 'required'
         ]);
-
+        //Handle File Upload
+        if($request->hasFile('cover_image')){
+            //Get filename with extencion.
+            $fileNameWithExt = $request ->file('cover_image')-> getClientOriginalName();
+            //Get just fileName
+            $fileName = pathInfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get just Ext
+            $fileExt = $request->file('cover_image')->getClientOriginalExtension();
+            //Filename to Store
+            $fileNameToStore = $fileName.'_'.time().'.'.$fileExt;
+            //Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
         //Update Post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
         return redirect('/posts')->with('success', 'Post Updated');
     }
@@ -104,6 +152,15 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        //Chek for correct user
+        if(auth()->user()->id != $post->user_id)
+        {
+           return redirect('/posts')->with('error', 'Unathorized');
+        }
+        //If image is not the default, deleats it
+        if($post->cover_image != 'noImage.jpg'){
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
     }
